Refactor TunerView into modular subviews
